move reduction scheduler to a separate file
